refactor: inclued pagination interface and updated pagination type object name
