+++ packages/challenges/src/modules/answers/typesObjects/PaginatedAnswers.typeObject.ts
@@ -1,12 +1,13 @@
 import { ObjectType, Field } from '@nestjs/graphql';
 import { Answer } from '../entities/Answer.entity';
-import { Pagination } from '../../challenges/typesObjects/pagination.typeObject';
+import { PaginationTypeObject } from '../../challenges/typesObjects/pagination.typeObject';
+import { Pagination } from 'src/core/domain/Pagination.interface';
 
 @ObjectType()
 export class PaginatedAnswers {
   @Field(() => [Answer])
   answers: Answer[];
 
-  @Field(() => Pagination)
+  @Field(() => PaginationTypeObject)
   pagination: Pagination;
 }
