+++ packages/challenges/src/modules/challenges/typesObjects/paginated-challenges.typeObject.ts
@@ -1,12 +1,13 @@
 import { ObjectType, Field } from '@nestjs/graphql';
+import { Pagination } from 'src/core/domain/Pagination.interface';
 import { Challenge } from '../entities/challenge.entity';
-import { Pagination } from './pagination.typeObject';
+import { PaginationTypeObject } from './pagination.typeObject';
 
 @ObjectType()
 export class PaginatedChallenges {
   @Field(() => [Challenge])
   challenges: Challenge[];
 
-  @Field(() => Pagination)
+  @Field(() => PaginationTypeObject)
   pagination: Pagination;
 }
